Add option to generate migrations suitable for running in a transaction
This adds a new PlanOpt WithTransactional which changes a few of the sql
generators so that they generate migration statements that can be run
within a transaction.

The notable ones that need to change are:
- CREATE INDEX CONCURRENTLY is now just CREATE INDEX
- Use ADD CONSTRAINT ... instead of ADD CONSTRAINT NOT VALID +
  VALIDATE CONSTRAINT.

Doing so required passing an option through a few different SQL vertex
generator types so there is now a generatorOptions struct that is used
to pass those on down.

This commit also introduces an new --transactional flag to the
pg-schema-diff command, which sets the WithTransactional planOpt.

Tests are included for the cases that have actually changed, along with
a test option to attempt to run the resulting DDL in a transaction.